% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{set_key_Renviron}
\alias{set_key_Renviron}
\title{Set authentication key in .Renviron}
\usage{
set_key_Renviron(name, key)
}
\arguments{
\item{name}{Name the key will refer to in .Renviron. Should be specific (ie HUBSPOT_KEY, not KEY)}

\item{key}{Authentication key - or other reference!}
}
\description{
Useful for accessing API keys outside of shared scripts. Key name chosen can be used as a reference on machine going forward.
}
\examples{
set_key_Renviron("PRODUCT_KEY", "mykey")
keyforscript <- Sys.getenv("PRODUCT_KEY")
}
