% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesforce.R
\name{sf_retrieve_ids}
\alias{sf_retrieve_ids}
\title{Retrieve Salesforce IDs from email}
\usage{
sf_retrieve_ids(object, email)
}
\arguments{
\item{object}{}

\item{email}{}
}
\description{
Retrieve Salesforce IDs from email
}
\examples{
sf_retrieve_ids("Contact", "email to lookup") 
email_list \%>\% pmap_df(sf_retrieve_ids)
}
